+++ CHANGELOG.md
@@ -38,6 +38,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Deprecated
 
 ### Removed
+- ONTOLOGY-179 Removed structured search components and utils
+- ONTOLOGY-179 Removed free text search components and utils
 - ONTOLOGY-156 Removed serialiseOwlFile util
 - ONTOLOGY-170 Removed required from rdfLabel in edit ontology form